Sacado:  More consolidation of Kokkos View files.
