--- packages/sacado/src/Kokkos_DynRankView_Fad.hpp
@@ -1,38 +1,0 @@
-// @HEADER
-// *****************************************************************************
-//                           Sacado Package
-//
-// Copyright 2006 NTESS and the Sacado contributors.
-// SPDX-License-Identifier: LGPL-2.1-or-later
-// *****************************************************************************
-// @HEADER
-
-#ifndef KOKKOS_DYN_RANK_VIEW_SACADO_FAD_HPP
-#define KOKKOS_DYN_RANK_VIEW_SACADO_FAD_HPP
-
-#include "Sacado_ConfigDefs.h"
-
-// This file is setup to always work even when KokkosContainers (which contains
-// Kokkos::DynRankView) isn't enabled.
-
-#if defined(HAVE_SACADO_KOKKOS)
-
-// We are hooking into Kokkos Core internals here
-// Need to define this macro since we include non-public headers
-#ifndef KOKKOS_IMPL_PUBLIC_INCLUDE
-#define KOKKOS_IMPL_PUBLIC_INCLUDE
-#define KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_CORE
-#endif
-// Only include forward declarations so any overloads appear before they
-// might be used inside Kokkos
-#include "Kokkos_Core_fwd.hpp"
-#include "Kokkos_Layout.hpp"
-//#include "Kokkos_DynRankView.hpp"
-#ifdef KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_CORE
-#undef KOKKOS_IMPL_PUBLIC_INCLUDE
-#undef KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_CORE
-#endif
-
-#endif // defined(HAVE_SACADO_KOKKOS)
-
-#endif /* #ifndef KOKKOS_DYN_RANK_VIEW_SACADO_FAD_HPP */
