--- packages/sacado/src/Kokkos_DynRankView_Fad_Contiguous.hpp
@@ -1,30 +1,0 @@
-// @HEADER
-// *****************************************************************************
-//                           Sacado Package
-//
-// Copyright 2006 NTESS and the Sacado contributors.
-// SPDX-License-Identifier: LGPL-2.1-or-later
-// *****************************************************************************
-// @HEADER
-
-#ifndef KOKKOS_DYN_RANK_VIEW_SACADO_FAD_CONTIGUOUS_HPP
-#define KOKKOS_DYN_RANK_VIEW_SACADO_FAD_CONTIGUOUS_HPP
-
-#include "Sacado_ConfigDefs.h"
-
-// This file is setup to always work even when KokkosContainers (which contains
-// Kokkos::DynRankView) isn't enabled.
-
-#if defined(HAVE_SACADO_KOKKOS)
-
-#include "Kokkos_DynRankView.hpp"
-
-#if defined(HAVE_SACADO_VIEW_SPEC) && !defined(SACADO_DISABLE_FAD_VIEW_SPEC)
-
-#include "Kokkos_View_Fad.hpp"
-
-#endif //defined(HAVE_SACADO_VIEW_SPEC) && !defined(SACADO_DISABLE_FAD_VIEW_SPEC)
-
-#endif // defined(HAVE_SACADO_KOKKOS)
-
-#endif /* #ifndef KOKKOS_DYN_RANK_VIEW_SACADO_FAD_HPP */
